broke samples into individual files
