+++ container_registry/container_analysis/src/sample/filter_vulnerability_occurrences.go
@@ -1,0 +1,70 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Samples for the Container Analysis golang libraries: https://cloud.google.com/container-registry/docs/container-analysis
+package sample
+
+// [START containeranalysis_filter_vulnerability_occurrences]
+
+import (
+	"context"
+	"fmt"
+	wait "k8s.io/apimachinery/pkg/util/wait"
+	"log"
+	"time"
+
+	containeranalysis "cloud.google.com/go/containeranalysis/apiv1beta1"
+	"google.golang.org/api/iterator"
+	discovery "google.golang.org/genproto/googleapis/devtools/containeranalysis/v1beta1/discovery"
+	grafeaspb "google.golang.org/genproto/googleapis/devtools/containeranalysis/v1beta1/grafeas"
+	vulnerability "google.golang.org/genproto/googleapis/devtools/containeranalysis/v1beta1/vulnerability"
+)
+
+// findHighSeverityVulnerabilitiesForImage retrieves a list of only high vulnerability occurrences associated with a resource.
+func findHighSeverityVulnerabilitiesForImage(resourceURL, projectID string) ([]*grafeaspb.Occurrence, error) {
+	// resourceURL := fmt.Sprintf("https://gcr.io/my-project/my-image")
+	ctx := context.Background()
+	client, err := containeranalysis.NewGrafeasV1Beta1Client(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("NewGrafeasV1Beta1Client: %v", err)
+	}
+	defer client.Close()
+
+	req := &grafeaspb.ListOccurrencesRequest{
+		Parent: fmt.Sprintf("projects/%s", projectID),
+		Filter: fmt.Sprintf("resourceUrl = %q kind = %q", resourceURL, "VULNERABILITY"),
+	}
+
+	var occurrenceList []*grafeaspb.Occurrence
+	it := client.ListOccurrences(ctx, req)
+	for {
+		occ, err := it.Next()
+		if err == iterator.Done {
+			break
+		}
+		if err != nil {
+			return nil, fmt.Errorf("occurrence iteration error: %v", err)
+		}
+
+		severityLevel := occ.GetVulnerability().GetSeverity()
+		if severityLevel == vulnerability.Severity_HIGH || severityLevel == vulnerability.Severity_CRITICAL {
+			occurrenceList = append(occurrenceList, occ)
+		}
+	}
+
+	return occurrenceList, nil
+}
+
+// [END containeranalysis_filter_vulnerability_occurrences]
+
