+++ container_registry/container_analysis/src/sample/get_discovery_info.go
@@ -1,0 +1,64 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Samples for the Container Analysis golang libraries: https://cloud.google.com/container-registry/docs/container-analysis
+package sample
+
+// [START containeranalysis_discovery_info]
+
+import (
+	"context"
+	"fmt"
+	"sync"
+	"time"
+
+	containeranalysis "cloud.google.com/go/containeranalysis/apiv1beta1"
+	pubsub "cloud.google.com/go/pubsub"
+	"google.golang.org/api/iterator"
+	grafeaspb "google.golang.org/genproto/googleapis/devtools/containeranalysis/v1beta1/grafeas"
+	"google.golang.org/genproto/googleapis/devtools/containeranalysis/v1beta1/vulnerability"
+)
+
+// getDiscoveryInfo retrieves and prints the Discovery Occurrence created for a specified image.
+// The Discovery Occurrence contains information about the initial scan on the image.
+func getDiscoveryInfo(w io.Writer, resourceURL, projectID string) error {
+	// resourceURL := fmt.Sprintf("https://gcr.io/my-project/my-image")
+	ctx := context.Background()
+	client, err := containeranalysis.NewGrafeasV1Beta1Client(ctx)
+	if err != nil {
+		return fmt.Errorf("NewGrafeasV1Beta1Client: %v", err)
+	}
+	defer client.Close()
+
+	req := &grafeaspb.ListOccurrencesRequest{
+		Parent: fmt.Sprintf("projects/%s", projectID),
+		Filter: fmt.Sprintf(`kind="DISCOVERY" AND resourceUrl=%q`, resourceURL),
+	}
+	it := client.ListOccurrences(ctx, req)
+	for {
+		occ, err := it.Next()
+		if err == iterator.Done {
+			break
+		}
+		if err != nil {
+			return fmt.Errorf("occurrence iteration error: %v", err)
+		}
+		fmt.Fprintln(w, occ)
+	}
+	return nil
+}
+
+// [END containeranalysis_discovery_info]
+
+
